Add boilerplate: model, controller, serializer
I'm starting to think that it might not be a good idea to keep these
microservices in separate codebases / apps. There is significant code
duplication happening among them, which isn't that big of a deal now,
but might cause problems if we decide to update our architecture, or
upgrade dependencies across the board.

A more sustainable solution might involve utilizing laravel-modules,
and treating each microservice as a package:

https://github.com/nWidart/laravel-modules

https://danluu.com/monorepo/