Align base controller with aggregator core
We were using a simplified base controller, but it didn't allow for
subresources. This change enables us to add subresource routes, e.g.
`publications/{id}/sections` and makes the microservice easier to
maintain, since its structure reflects that of our primary codebase.

Part of the problem is that Lumen's support for retrieving route
params is lacking. Laravel and Lumen share the same Request class,
but use different routing libraries. The base class doesn't account
for that fact in this specific regard. The situation is unlikely to
be fixed anytime soon. See the following threads, among others:

https://github.com/laravel/lumen-framework/issues/119
https://github.com/laravel/framework/pull/9417

We thus have to retrieve route params in a somewhat hacky way:

    $id = $request->route()[2]['id'] ?? null;

If the issue is ever resolved upstream, we can change it to this:

    $id = $request->route('id');

Since using Lumen has presented similar issues in the past, it might
be best to use Laravel and manually strip down its bootstrap process
to the minimum of what we need. It won't be as fast, but since this
microservice is not likely to see significant load, this is not an
issue. Alternatively, we can migrate this code into the core of the
aggregator to make maintainance easier.