Add accession parsing to section model
We currently do this for each section, for each request. This isn't
efficient or speedy at all, comparatively, but it's good for a proof
of concept. I'd like to run some analytics to see what numbers to
expect, before moving these calls to the import script.

I've enabled Facades to avoid `A facade root has not been set.` error
in `Section->getContent()`. There might be another way to fix it.