Copy configs from data-service-library
Laravel doesn't have a mechanism for default config fallbacks.
Until we can resolve this in the foundation, these need to be
copied to each Laravel-based data service.

Note that `app.php` isn't default, i.e. different from how it
was generated by Laravel. Lots of things were commented out in
an effort to slim-down the app.