Update foundation to use its RouteServiceProvider
All of the API-only microservices use the same route scheme. To reduce
duplication across our repos and to ease their migration from Lumen to
Laravel, I've included a default `RouteServiceProvider.php` in the
foundation. All it does it load `routes/api.php`.

Please note that you can manually create a provider in this application
codebase, and modify `app.php` to use it instead. Depending on how API
versioning is going to work out in the future, per-service providers
might be the way to go.